+++ AlchemistCode/Cards/Common/Vitrify.cs
@@ -14,7 +14,7 @@ public class Vitrify : AlchemistCard
         WithTip(typeof(PlatingPower));
     }
 
-    protected override bool ConditionalGlow => IsEnchanted;
+    internal override bool GainsEffectWhenEnchanted => true;
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {