+++ AlchemistCode/Cards/Uncommon/Dissect.cs
@@ -15,7 +15,7 @@ public class Dissect : AlchemistCard
         WithTip(typeof(VulnerablePower));
     }
 
-    protected override bool ConditionalGlow => IsEnchanted;
+    internal override bool GainsEffectWhenEnchanted => true;
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {